+++ src/xSTIR/cSTIR/include/sirf/STIR/cstir.h
@@ -89,6 +89,11 @@ extern "C" {
 	void* cSTIR_acquisitionModelBwdReplace(void* ptr_am, void* ptr_ad,
 		int subset_num, int num_subsets, void* ptr_im);
 
+        // Acquisition Model Matrix
+        void* cSTIR_SPECTUBMatrixSetResolution
+                 (const void* ptr_acq_matrix,
+                  const float collimator_sigma_0_in_mm, const float collimator_slope_in_mm, const bool full_3D);
+
 	// Acquisition data methods
 	void* cSTIR_getAcquisitionDataStorageScheme();
 	void* cSTIR_setAcquisitionDataStorageScheme(const char* scheme);