+++ CHANGES.md
@@ -12,6 +12,8 @@
   - (C++) Replaced where possible returning `stir::Succeeded::no` with throwing exception.
   - (C++) Fixed a bug in `PETAcquisitionDataInMemory::norm`.
 
+* added SPECTUBMatrix for (simple) usage in SPECT
+
 * SIRF Python interface
   - Removed `__div__` ,  `__idiv__` operators for `DataContainers` required for Python2.
   - Added `__truediv__` and `__itruediv__` Python3 operators to DataContainer algebra.